refactor: 이전 버전의 코드를 다른 패키지로 분리한다.
