+++ android/app/src/main/AndroidManifest.xml
@@ -18,9 +18,6 @@
         tools:targetApi="31">
         <activity
             android:name=".feature.home.HomeActivity"
-            android:exported="false" />
-        <activity
-            android:name=".feature.statsearching.StatsActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -28,6 +25,10 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+
+        <activity
+            android:name=".feature.statsearching.legacy.StatsActivity"
+            android:exported="false" />
 
         <meta-data
             android:name="preloaded_fonts"
