+++ android/app/src/main/java/com/foss/foss/di/manual/ViewModelFactoryProvider.kt
@@ -3,7 +3,7 @@ package com.foss.foss.di.manual
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.viewmodel.initializer
 import androidx.lifecycle.viewmodel.viewModelFactory
-import com.foss.foss.feature.statsearching.StatsViewModel
+import com.foss.foss.feature.statsearching.legacy.StatsViewModel
 
 object ViewModelFactoryProvider {
