+++ android/app/src/main/res/layout/activity_stats.xml
@@ -6,7 +6,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context="com.foss.foss.feature.statsearching.StatsActivity">
+        tools:context="com.foss.foss.feature.statsearching.legacy.StatsActivity">
 
         <EditText
             android:id="@+id/stats_et_nickname"
