+++ android/app/src/main/res/layout/fragment_recent_stats.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="#272D31"
+    tools:context=".feature.statsearching.recent.RecentStatsFragment">
+
+    <!-- TODO: Update blank fragment layout -->
+    <TextView
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:text="최근 전적 fragment"
+        android:textColor="@color/white"
+        android:textSize="50sp" />
+
+</FrameLayout>
