+++ android/app/src/main/res/layout/fragment_relative_stats.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:background="#272D31"
+    tools:context=".feature.statsearching.relative.RelativeStatsFragment">
+
+    <TextView
+        android:textSize="50sp"
+        android:textColor="@color/white"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:text="상대전적 fragment" />
+
+</FrameLayout>
